FIX: functional test for importers
