+++ importers/onesite/Makefile
@@ -2,11 +2,15 @@ onesite: *.go
 	go vet
 	golint
 	go build
+	go test
+	go install
 
 clean:
+	go clean
 	@echo "*** CLEAN completed in importers/onesite ***"
 
 test:
+	go test
 	@echo "*** TEST completed in importers/onesite ***"
 
 #man:
@@ -14,6 +18,4 @@ test:
 #	cp importers/onesite.1 /usr/local/share/man/man1
 
 package:
-	mkdir -p ../../tmp/rentroll/importers/onesite/
-	cp mapper.json ../../tmp/rentroll/importers/onesite/
 	@echo "*** PACKAGE completed in importers/onesite ***"
