+++ test/importers/Makefile
@@ -2,16 +2,16 @@ DIRS = onesite
 
 importers:
 	for dir in $(DIRS); do make -C $$dir;done
-	@echo "*** Completed in importers ***"
+	@echo "*** Completed in test/importers ***"
 
 clean:
 	for dir in $(DIRS); do make -C $$dir clean;done
-	@echo "*** CLEAN completed in importers ***"
+	@echo "*** CLEAN completed in test/importers ***"
 
 test:
 	for dir in $(DIRS); do make -C $$dir test;done
-	@echo "*** TEST completed in importers ***"
+	@echo "*** TEST completed in test/importers ***"
 
 package:
 	for dir in $(DIRS); do make -C $$dir package;done
-	@echo "*** PACKAGING COMPLETE IN importers ***"
+	@echo "*** PACKAGING COMPLETE IN test/importers ***"
