+++ test/importers/onesite/Makefile
@@ -1,14 +1,14 @@
 onesite:
-	@echo "*** Completed in onesite ***"
+	@echo "*** Completed in test/importers/onesite ***"
 
 clean:
-	TEMPCSVSTORE="../../../importers/onesite/tempCSVs"
-	rm -f [a-z] rentroll.log log *.txt ${TEMPCSVSTORE}/*.csv rt.csv people.csv rentable.csv ra.csv custom.csv
-	@echo "*** CLEAN completed in onesite ***"
+	TEMPCSVSTORE="../../../importers/onesite/temp_CSVs"
+	rm -f [a-z] rentroll.log log *.txt ${TEMPCSVSTORE}/*.csv
+	@echo "*** CLEAN completed in test/importers/onesite ***"
 
 test:
 	./functest.sh
-	@echo "*** TEST completed in onesite ***"
+	@echo "*** TEST completed in test/importers/onesite ***"
 
 package:
-	@echo "*** PACKAGE completed in onesite ***"
+	@echo "*** PACKAGE completed in test/importers/onesite ***"
