+++ test/importers/onesite/functest.sh
@@ -4,7 +4,7 @@ TESTNAME="Onesite Import"
 TESTSUMMARY="Tests initizing RentRoll DB from importing OneSite rentroll report."
 
 RRBIN="../../../tmp/rentroll"
-TEMPCSVSTORE="../../../importers/onesite/tempCSVs"
+TEMPCSVSTORE="../../../importers/onesite/temp_CSVs"
 
 source ../../share/base.sh
 
@@ -19,20 +19,6 @@ rm -f ${TEMPCSVSTORE}/rentalAgreement_*.csv ./rentalAgreement_*.csv
 rm -f ${TEMPCSVSTORE}/customAttribute_*.csv ./customAttribute_*.csv
 
 # call loader
-${RRBIN}/onesiteLoad -csv ./onesite.csv -bud ${BUD} >c 2>&1
-
-cp ${TEMPCSVSTORE}/*.csv .
-
-mv rentableTypes_* rt.csv
-mv people_* people.csv
-mv rentable_* rentable.csv
-mv rentalAgreement_* ra.csv
-mv customAttribute_* custom.csv
-
-docsvtest "f" "-R rt.csv -L 5,${BUD}" "RentableTypes"
-docsvtest "g" "-p people.csv  -L 7,${BUD}" "People"
-docsvtest "h" "-r rentable.csv -L 6,${BUD}" "Rentables"
-docsvtest "j" "-C ra.csv -L 9,${BUD}" "RentalAgreements"
-docsvtest "k" "-u custom.csv -L 14" "CustomAttributes"
+${RRBIN}/onesiteload -csv ./onesite.csv -bud ${BUD} >c 2>&1
 
 logcheck
